+++ src/editor/codemirror/CodeMirror.tsx
@@ -6,12 +6,13 @@
 import { highlightActiveLineGutter, lineNumbers } from "@codemirror/gutter";
 import { redoDepth, undoDepth } from "@codemirror/history";
 import { EditorSelection, EditorState, Extension } from "@codemirror/state";
-import { EditorView, highlightActiveLine } from "@codemirror/view";
+import { EditorView, highlightActiveLine, ViewUpdate } from "@codemirror/view";
 import { useEffect, useMemo, useRef } from "react";
 import { useIntl } from "react-intl";
-import { lineNumFromUint8Array } from "../../fs/fs";
+import { lineNumFromUint8Array } from "../../common/text-util";
 import { createUri } from "../../language-server/client";
 import { useLanguageServerClient } from "../../language-server/language-server-hooks";
+import { Logging } from "../../logging/logging";
 import { useLogging } from "../../logging/logging-hooks";
 import { useRouterState } from "../../router-hooks";
 import { ParameterHelpOption } from "../../settings/settings";
@@ -94,19 +95,7 @@ const CodeMirror = ({
             undo: undoDepth(view.state),
             redo: redoDepth(view.state),
           });
-          // Log the number of lines pasted by the user. Do not count
-          // empty lines before and after the pasted content.
-          if (update.transactions[0].isUserEvent("input.paste")) {
-            update.changes.iterChangedRanges((_fromA, _toA, fromB, toB) => {
-              const pastedContent = update.state.sliceDoc(fromB, toB).trim();
-              logging.event({
-                type: "user-paste",
-                value: lineNumFromUint8Array(
-                  new TextEncoder().encode(pastedContent)
-                ),
-              });
-            });
-          }
+          logPastedLineCount(logging, update);
         }
       });
       const state = EditorState.create({
@@ -239,5 +228,24 @@ const CodeMirror = ({
 function themeExtensionsForOptions(options: { fontSize: number }): Extension {
   return themeExtensions(options.fontSize + "pt");
 }
+
+const logPastedLineCount = (logging: Logging, update: ViewUpdate) => {
+  update.transactions
+    .filter((transaction) => transaction.isUserEvent("input.paste"))
+    .forEach((transaction) =>
+      transaction.changes.iterChanges(
+        (_fromA, _toA, _fromB, _toB, inserted) => {
+          const lineCount = lineNumFromUint8Array(
+            // Ignore leading/trailing lines.
+            new TextEncoder().encode(inserted.toString().trim())
+          );
+          logging.event({
+            type: "paste",
+            value: lineCount,
+          });
+        }
+      )
+    );
+};
 
 export default CodeMirror;
