+++ src/fs/fs.ts
@@ -23,6 +23,7 @@ import {
   SplitStrategyStorage,
 } from "./storage";
 import { PythonProject } from "./initial-project";
+import { lineNumFromUint8Array } from "../common/text-util";
 
 const commonFsSize = 20 * 1024;
 
@@ -131,20 +132,6 @@ export const diff = (before: Project, after: Project): FileChange[] => {
     }
   }
   return result;
-};
-
-export const lineNumFromUint8Array = (arr: Uint8Array): number => {
-  let lineCount = 1;
-  let prevByte: number | undefined;
-  const LF = 10;
-  const CR = 13;
-  arr.forEach((byte) => {
-    if ((byte === LF && prevByte !== CR) || byte === CR) {
-      lineCount++;
-    }
-    prevByte = byte;
-  });
-  return lineCount;
 };
 
 export const EVENT_PROJECT_UPDATED = "project_updated";
